chore (refs T22614): move procedure lib folder to client/js.
