+++ client/js/lib/procedure/charts/BarChart.js
