+++ client/js/lib/procedure/charts/BarPieChart.js
