+++ client/js/lib/procedure/charts/Legend.js
