+++ client/js/lib/procedure/charts/helpers/getColorsFromCSS.js
